[Gutenberg] Moved stockPhotos logic to its own utils class
